backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
librosa==0.10.1
matplotlib==3.7.2
numpy==1.24.3
Pillow==10.0.0
scipy==1.11.1

